refactor: breakout message input logic to own file
